+++ src/OracleMiddleware/OracleMiddleware.sol
@@ -52,7 +52,7 @@ contract OracleMiddleware is IOracleMiddleware, PythOracle, ChainlinkOracle, Own
     { }
 
     /* -------------------------------------------------------------------------- */
-    /*                          Price retrieval features                          */
+    /*                           Public view functions                            */
     /* -------------------------------------------------------------------------- */
 
     /// @inheritdoc IOracleMiddleware
@@ -63,46 +63,42 @@ contract OracleMiddleware is IOracleMiddleware, PythOracle, ChainlinkOracle, Own
         returns (PriceInfo memory price_)
     {
         if (action == ProtocolAction.None) {
-            return _getPythOrChainlinkDataStreamPrice(data, targetTimestamp, ConfidenceInterval.None);
+            return _getLowLatencyPrice(data, targetTimestamp, ConfidenceInterval.None);
         } else if (action == ProtocolAction.Initialize) {
-            return _getOnChainPrice(data, targetTimestamp, ConfidenceInterval.None);
+            return _getInitiateActionPrice(data, ConfidenceInterval.None);
         } else if (action == ProtocolAction.ValidateDeposit) {
             // Use the lowest price in the confidence interval to ensure a minimum benefit for the user in case
             // of price inaccuracies
-            return _getPythOrChainlinkDataStreamPrice(data, targetTimestamp, ConfidenceInterval.Down);
+            return _getLowLatencyPrice(data, targetTimestamp, ConfidenceInterval.Down);
         } else if (action == ProtocolAction.ValidateWithdrawal) {
             // Use the highest price in the confidence interval to ensure a minimum benefit for the user in case
             // of price inaccuracies
-            return _getPythOrChainlinkDataStreamPrice(data, targetTimestamp, ConfidenceInterval.Up);
+            return _getLowLatencyPrice(data, targetTimestamp, ConfidenceInterval.Up);
         } else if (action == ProtocolAction.ValidateOpenPosition) {
             // Use the highest price in the confidence interval to ensure a minimum benefit for the user in case
             // of price inaccuracies
-            return _getPythOrChainlinkDataStreamPrice(data, targetTimestamp, ConfidenceInterval.Up);
+            return _getLowLatencyPrice(data, targetTimestamp, ConfidenceInterval.Up);
         } else if (action == ProtocolAction.ValidateClosePosition) {
             // Use the lowest price in the confidence interval to ensure a minimum benefit for the user in case
             // of price inaccuracies
-            return _getPythOrChainlinkDataStreamPrice(data, targetTimestamp, ConfidenceInterval.Down);
+            return _getLowLatencyPrice(data, targetTimestamp, ConfidenceInterval.Down);
         } else if (action == ProtocolAction.Liquidation) {
             // Special case, if we pass a timestamp of zero, then we accept all prices newer than `_recentPriceDelay`
-            return _getPythOrChainlinkDataStreamPrice(data, 0, ConfidenceInterval.None);
+            return _getLowLatencyPrice(data, 0, ConfidenceInterval.None);
         } else if (action == ProtocolAction.InitiateDeposit) {
-            // If the user chooses to intitiate with a pyth price, we apply the relevant confidence interval adjustment
-            return _getOnChainPrice(data, targetTimestamp, ConfidenceInterval.Down);
+            // If the user chooses to initiate with a pyth price, we apply the relevant confidence interval adjustment
+            return _getInitiateActionPrice(data, ConfidenceInterval.Down);
         } else if (action == ProtocolAction.InitiateWithdrawal) {
-            // If the user chooses to intitiate with a pyth price, we apply the relevant confidence interval adjustment
-            return _getOnChainPrice(data, targetTimestamp, ConfidenceInterval.Up);
+            // If the user chooses to initiate with a pyth price, we apply the relevant confidence interval adjustment
+            return _getInitiateActionPrice(data, ConfidenceInterval.Up);
         } else if (action == ProtocolAction.InitiateOpenPosition) {
-            // If the user chooses to intitiate with a pyth price, we apply the relevant confidence interval adjustment
-            return _getOnChainPrice(data, targetTimestamp, ConfidenceInterval.Up);
+            // If the user chooses to initiate with a pyth price, we apply the relevant confidence interval adjustment
+            return _getInitiateActionPrice(data, ConfidenceInterval.Up);
         } else if (action == ProtocolAction.InitiateClosePosition) {
-            // If the user chooses to intitiate with a pyth price, we apply the relevant confidence interval adjustment
-            return _getOnChainPrice(data, targetTimestamp, ConfidenceInterval.Down);
+            // If the user chooses to initiate with a pyth price, we apply the relevant confidence interval adjustment
+            return _getInitiateActionPrice(data, ConfidenceInterval.Down);
         }
     }
-
-    /* -------------------------------------------------------------------------- */
-    /*                              Generic features                              */
-    /* -------------------------------------------------------------------------- */
 
     /// @inheritdoc IOracleMiddleware
     function getValidationDelay() external view returns (uint256) {
@@ -115,14 +111,137 @@ contract OracleMiddleware is IOracleMiddleware, PythOracle, ChainlinkOracle, Own
     }
 
     /// @inheritdoc IOracleMiddleware
-    function validationCost(bytes calldata data, ProtocolAction) external view virtual returns (uint256 result_) {
+    function getMaxConfRatio() external pure returns (uint16) {
+        return MAX_CONF_RATIO;
+    }
+
+    /// @inheritdoc IOracleMiddleware
+    function getConfRatioDenom() external pure returns (uint16) {
+        return CONF_RATIO_DENOM;
+    }
+
+    /// @inheritdoc IOracleMiddleware
+    function getConfRatio() external view returns (uint16) {
+        return _confRatio;
+    }
+
+    /// @inheritdoc IOracleMiddleware
+    function validationCost(bytes calldata data, ProtocolAction) public view virtual returns (uint256 result_) {
         if (data.length > 0) {
             result_ = _getPythUpdateFee(data);
         }
     }
 
     /* -------------------------------------------------------------------------- */
-    /*                               Owner features                               */
+    /*                             Internal functions                             */
+    /* -------------------------------------------------------------------------- */
+
+    /**
+     * @dev Get the price from the low-latency oracle (at the moment only Pyth, later maybe others might be supported).
+     * @param data The signed price update data.
+     * @param actionTimestamp The timestamp of the action corresponding to the price. If zero, then we must accept all
+     * recent prices according to `_recentPriceDelay`.
+     * @param dir The direction for the confidence interval adjusted price.
+     * @return price_ The price from the low-latency oracle, adjusted according to the confidence interval direction.
+     */
+    function _getLowLatencyPrice(bytes calldata data, uint128 actionTimestamp, ConfidenceInterval dir)
+        internal
+        returns (PriceInfo memory price_)
+    {
+        // If actionTimestamp is 0 we're performing a liquidation and we don't add the validation delay
+        if (actionTimestamp > 0) {
+            // Add the validation delay to the action timestamp to get the timestamp of the price data used to
+            // validate
+            actionTimestamp += uint128(_validationDelay);
+        }
+        FormattedPythPrice memory pythPrice = _getFormattedPythPrice(data, actionTimestamp, MIDDLEWARE_DECIMALS);
+
+        price_ = _adjustPythPrice(pythPrice, dir);
+    }
+
+    /**
+     * @notice Get the price for an initiate action of the protocol.
+     * @dev If the data parameter is not empty, validate the price with PythOracle. Else, get the on-chain price from
+     * Chainlink and compare its timestamp with the latest seen Pyth price (cached). If Pyth is more recent, we return
+     * it. Otherwise we return the chainlink price. In case of chainlink price, we don't have a confidence interval and
+     * so both `neutralPrice` and `price` are equal.
+     * @param data An optional VAA from Pyth
+     * @param dir The direction for applying the confidence interval (in case we use a Pyth price)
+     * @return price_ The price to use for the user action
+     */
+    function _getInitiateActionPrice(bytes calldata data, ConfidenceInterval dir)
+        internal
+        returns (PriceInfo memory price_)
+    {
+        // If data is not empty, use pyth
+        if (data.length > 0) {
+            // since we use this function for `initiate` type actions which pass `targetTimestamp = block.timestamp`,
+            // we should pass `0` to the function below to signal that we accept any recent price
+            return _getLowLatencyPrice(data, 0, dir);
+        }
+
+        ChainlinkPriceInfo memory chainlinkOnChainPrice = _getFormattedChainlinkPrice(MIDDLEWARE_DECIMALS);
+
+        // check if the cached pyth price is more recent and return it instead
+        FormattedPythPrice memory latestPythPrice = _getLatestStoredPythPrice(MIDDLEWARE_DECIMALS);
+        if (chainlinkOnChainPrice.timestamp <= latestPythPrice.publishTime) {
+            // We use the same price age limit as for chainlink here
+            if (latestPythPrice.publishTime < block.timestamp - _timeElapsedLimit) {
+                revert OracleMiddlewarePriceTooOld(latestPythPrice.publishTime);
+            }
+            return _adjustPythPrice(latestPythPrice, dir);
+        }
+
+        // If the price equals PRICE_TOO_OLD then the tolerated time elapsed for price validity was exceeded, revert.
+        if (chainlinkOnChainPrice.price == PRICE_TOO_OLD) {
+            revert OracleMiddlewarePriceTooOld(chainlinkOnChainPrice.timestamp);
+        }
+
+        // If the price is negative or zero, revert.
+        if (chainlinkOnChainPrice.price <= 0) {
+            revert OracleMiddlewareWrongPrice(chainlinkOnChainPrice.price);
+        }
+
+        price_ = PriceInfo({
+            price: uint256(chainlinkOnChainPrice.price),
+            neutralPrice: uint256(chainlinkOnChainPrice.price),
+            timestamp: chainlinkOnChainPrice.timestamp
+        });
+    }
+
+    /**
+     * @notice Apply the confidence interval in the `dir` direction, scaling it by the configured `_confRatio`.
+     * @param pythPrice The formatted Pyth price object.
+     * @param dir The direction to apply the confidence interval.
+     */
+    function _adjustPythPrice(FormattedPythPrice memory pythPrice, ConfidenceInterval dir)
+        internal
+        view
+        returns (PriceInfo memory price_)
+    {
+        if (dir == ConfidenceInterval.Down) {
+            uint256 adjust = (pythPrice.conf * _confRatio) / CONF_RATIO_DENOM;
+            if (adjust >= pythPrice.price) {
+                // avoid underflow or zero price due to confidence interval adjustment
+                price_.price = 1;
+            } else {
+                // strictly positive
+                unchecked {
+                    price_.price = pythPrice.price - adjust;
+                }
+            }
+        } else if (dir == ConfidenceInterval.Up) {
+            price_.price = pythPrice.price + ((pythPrice.conf * _confRatio) / CONF_RATIO_DENOM);
+        } else {
+            price_.price = pythPrice.price;
+        }
+
+        price_.timestamp = pythPrice.publishTime;
+        price_.neutralPrice = pythPrice.price;
+    }
+
+    /* -------------------------------------------------------------------------- */
+    /*                            Privileged functions                            */
     /* -------------------------------------------------------------------------- */
 
     /// @inheritdoc IOracleMiddleware
@@ -150,98 +269,6 @@ contract OracleMiddleware is IOracleMiddleware, PythOracle, ChainlinkOracle, Own
         _recentPriceDelay = newDelay;
 
         emit RecentPriceDelayUpdated(newDelay);
-    }
-
-    /* -------------------------------------------------------------------------- */
-    /*                    Specialized price retrieval methods                     */
-    /* -------------------------------------------------------------------------- */
-
-    /**
-     * @dev Get the price from Pyth or Chainlink, depending on the data.
-     * @param data The data used to get the price.
-     * @param actionTimestamp The timestamp of the action corresponding to the price. If zero, then we must accept all
-     * recent prices.
-     * @param conf The confidence interval to use.
-     */
-    function _getPythOrChainlinkDataStreamPrice(bytes calldata data, uint128 actionTimestamp, ConfidenceInterval conf)
-        internal
-        returns (PriceInfo memory price_)
-    {
-        // If actionTimestamp is 0 we're performing a liquidation and we don't add the validation delay
-        if (actionTimestamp > 0) {
-            actionTimestamp += uint128(_validationDelay);
-        }
-        // Add the validation delay to the action timestamp to get the timestamp of the price data used to
-        // validate
-        FormattedPythPrice memory pythPrice = _getFormattedPythPrice(data, actionTimestamp, MIDDLEWARE_DECIMALS);
-
-        if (conf == ConfidenceInterval.Down) {
-            uint256 adjust = (pythPrice.conf * _confRatio) / CONF_RATIO_DENOM;
-            if (adjust >= pythPrice.price) {
-                // avoid underflow or zero price due to confidence interval adjustment
-                price_.price = 1;
-            } else {
-                // strictly positive
-                unchecked {
-                    price_.price = pythPrice.price - adjust;
-                }
-            }
-        } else if (conf == ConfidenceInterval.Up) {
-            price_.price = pythPrice.price + ((pythPrice.conf * _confRatio) / CONF_RATIO_DENOM);
-        } else {
-            price_.price = pythPrice.price;
-        }
-
-        price_.timestamp = pythPrice.publishTime;
-        price_.neutralPrice = pythPrice.price;
-    }
-
-    /// @dev If the data parameter is not empty, get the price from pyth, else, get it from chainlink.
-    function _getOnChainPrice(bytes calldata data, uint128 actionTimestamp, ConfidenceInterval conf)
-        internal
-        returns (PriceInfo memory)
-    {
-        // If data is not empty, use pyth
-        if (data.length > 0) {
-            return _getPythOrChainlinkDataStreamPrice(data, actionTimestamp, conf);
-        }
-
-        ChainlinkPriceInfo memory chainlinkOnChainPrice = _getFormattedChainlinkPrice(MIDDLEWARE_DECIMALS);
-
-        // If the price equals PRICE_TOO_OLD then the tolerated time elapsed for price validity was exceeded, revert.
-        if (chainlinkOnChainPrice.price == PRICE_TOO_OLD) {
-            revert OracleMiddlewarePriceTooOld(chainlinkOnChainPrice.timestamp);
-        }
-
-        // If the price is negative or zero, revert.
-        if (chainlinkOnChainPrice.price <= 0) {
-            revert OracleMiddlewareWrongPrice(chainlinkOnChainPrice.price);
-        }
-
-        return PriceInfo({
-            price: uint256(chainlinkOnChainPrice.price),
-            neutralPrice: uint256(chainlinkOnChainPrice.price),
-            timestamp: chainlinkOnChainPrice.timestamp
-        });
-    }
-
-    /* -------------------------------------------------------------------------- */
-    /*                           Price confidence ratio                           */
-    /* -------------------------------------------------------------------------- */
-
-    /// @inheritdoc IOracleMiddleware
-    function getMaxConfRatio() external pure returns (uint16) {
-        return MAX_CONF_RATIO;
-    }
-
-    /// @inheritdoc IOracleMiddleware
-    function getConfRatioDenom() external pure returns (uint16) {
-        return CONF_RATIO_DENOM;
-    }
-
-    /// @inheritdoc IOracleMiddleware
-    function getConfRatio() external view returns (uint16) {
-        return _confRatio;
     }
 
     /// @inheritdoc IOracleMiddleware