+++ src/interfaces/UsdnProtocol/IUsdnProtocolErrors.sol
@@ -93,11 +93,32 @@ interface IUsdnProtocolErrors {
     /// @dev Indicates that the provided address for the LiquidationRewardsManager contract address is invalid.
     error UsdnProtocolInvalidLiquidationRewardsManagerAddress();
 
-    /// @dev Indicates that the provided fee percentage value is invalid.
+    /// @dev Indicates that the provided fee basis point value is invalid.
     error UsdnProtocolInvalidProtocolFeeBps();
 
     /// @dev Indicates that the provided fee collector address is invalid
     error UsdnProtocolInvalidFeeCollector();
+
+    /// @dev Indicates that the provided security deposit is lower than _securityDepositValue
+    error UsdnProtocolSecurityDepositTooLow();
+
+    /// @dev Indicates that the wanted security deposit value cannot be set
+    error UsdnProtocolInvalidSecurityDepositValue();
+
+    /// @dev Indicates that the ether balance of the contract at the end of the action is not the expected one
+    error UsdnProtocolUnexpectedBalance();
+
+    /// @dev Indicates that the soft longExpo imbalance limit provided is invalid
+    error UsdnProtocolInvalidExpoImbalanceLimit();
+
+    /// @dev Indicates that the protocol imbalance limit is reached
+    error UsdnProtocolImbalanceLimitReached(int256 imbalanceBps);
+
+    /// @dev Indicates that the protocol vault expo is invalid
+    error UsdnProtocolInvalidVaultExpo();
+
+    /// @dev Indicates that the protocol long expo is invalid
+    error UsdnProtocolInvalidLongExpo();
 
     /**
      * @dev Indicates that the data provided to validate an actionable pending action is invalid (zero length or length