+++ test/unit/UsdnProtocol/Long/Long.t.sol
@@ -118,21 +118,19 @@ contract TestUsdnProtocolLong is UsdnProtocolBaseFixture {
      * @custom:and a leverage of 2x (liquidation price $500)
      * @custom:or a leverage of 4x (liquidation price $750)
      * @custom:when The current price is $2000 and the leverage is 2x
-     * @custom:or the current price is $1000 and the leverage is 2x
-     * @custom:or the current price is $500 and the leverage is 2x
-     * @custom:or the current price is $2000 and the leverage is 4x
-     * @custom:then The position value is 1.5 wstETH ($2000 at 2x)
-     * @custom:or the position value is 1 wstETH ($1000 at 2x)
-     * @custom:or the position value is 0 wstETH ($500 at 2x)
-     * @custom:or the position value is 2.5 wstETH ($2000 at 4x)
+     * @custom:then The position value is 1.5 wstETH
+     * @custom:when the current price is $1000 and the leverage is 2x
+     * @custom:then the position value is 1 wstETH
+     * @custom:when the current price is $500 and the leverage is 2x
+     * @custom:then the position value is 0 wstETH
+     * @custom:when the current price is $200 and the leverage is 2x
+     * @custom:then the position value is -3 wstETH
+     * @custom:when the current price is $2000 and the leverage is 4x
+     * @custom:then the position value is 2.5 wstETH
      */
     function test_positionValue() public {
-        uint128 positionTotalExpo = uint128(
-            FixedPointMathLib.fullMulDiv(
-                1 ether, 2 * 10 ** protocol.LEVERAGE_DECIMALS(), 10 ** protocol.LEVERAGE_DECIMALS()
-            )
-        );
-        uint256 value = protocol.i_positionValue(2000 ether, 500 ether, positionTotalExpo);
+        uint128 positionTotalExpo = 2 ether;
+        int256 value = protocol.i_positionValue(2000 ether, 500 ether, positionTotalExpo);
         assertEq(value, 1.5 ether, "Position value should be 1.5 ether");
 
         value = protocol.i_positionValue(1000 ether, 500 ether, positionTotalExpo);
@@ -141,11 +139,10 @@ contract TestUsdnProtocolLong is UsdnProtocolBaseFixture {
         value = protocol.i_positionValue(500 ether, 500 ether, positionTotalExpo);
         assertEq(value, 0 ether, "Position value should be 0");
 
-        positionTotalExpo = uint128(
-            FixedPointMathLib.fullMulDiv(
-                1 ether, 4 * 10 ** protocol.LEVERAGE_DECIMALS(), 10 ** protocol.LEVERAGE_DECIMALS()
-            )
-        );
+        value = protocol.i_positionValue(200 ether, 500 ether, positionTotalExpo);
+        assertEq(value, -3 ether, "Position value should be negative");
+
+        positionTotalExpo = 4 ether;
         value = protocol.i_positionValue(2000 ether, 750 ether, positionTotalExpo);
         assertEq(value, 2.5 ether, "Position with 4x leverage should have a 2.5 ether value");
     }